app/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8.5,<3.0
django-extensions>=3.0.5
djangorestframework>=3.12.2
djangorestframework-simplejwt>=4.4.0
